requirements.txt
discord.py==1.4.1
discord==1.0.1
python-dotenv==0.15.0
pynacl

sympy~=1.9
pyyaml~=5.4.1

numpy~=1.21.2

pydub~=0.25.1

GitPython
